+++ test/osutils.jl
@@ -40,8 +40,8 @@ if Sys.iswindows()
     @testset "path variables use correct path delimiters on windows" begin
         for path in (Base.SYSCONFDIR, Base.DATAROOTDIR, Base.DOCDIR,
                      Base.LIBDIR, Base.PRIVATE_LIBDIR, Base.INCLUDEDIR)
-            @test !contains(path, "/")
-            @test !contains(path, "\\\\")
+            @test !isfound("/", path)
+            @test !isfound("\\\\", path)
         end
     end
 end