+++ test/strings/search.jl
@@ -314,9 +314,9 @@ end
 @test findfirst(r"az", "foo,bar,baz") == 10:11
 @test findnext(r"az", "foo,bar,baz", 12) == nothing
 
-# contains with a String and Char needle
-@test contains("foo", "o")
-@test contains("foo", 'o')
+# isfound with a String and Char needle
+@test isfound("o", "foo")
+@test isfound('o', "foo")
 
 @test_throws ErrorException "ab" ∈ "abc"